backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
torch==2.1.2
transformers==4.35.2
python-jose==3.3.0
pydantic==2.5.2
numpy==1.24.3
requests>=2.31.0 